+++ include/mainwindow.h
@@ -25,6 +25,7 @@ class MainWindow : public QMainWindow
 public:
     explicit MainWindow(QWidget *parent = 0);
     ~MainWindow();
+    virtual void closeEvent(QCloseEvent *event);
 
 private:
     Ui::MainWindow *ui;
@@ -34,8 +35,10 @@ private:
     QTranslator m_qtTranslator;
     QTranslator m_translator;
     QSettings *m_settings;
+    QString m_language;
 
 public slots:
+    void refreshTitleBar(bool modified);
     void refreshPreview();
     void refreshScenesView();
     void exportAsPDF();
@@ -57,6 +60,7 @@ private slots:
     void slot_actionAbout_Magic_Fountain();
 
     void slot_clickScenes(QListWidgetItem *item);
+    void slot_checkAndSaveScript();
 
 protected:
     void changeEvent(QEvent *event);