+++ src/mainwindow.cpp
@@ -8,7 +8,7 @@
 
 #include "block.h"
 
-QString GLOBAL_VERSION = "1.0.0";
+QString GLOBAL_VERSION = "1.0.0-alpha";
 
 MainWindow::MainWindow(QWidget *parent) :
     QMainWindow(parent),
@@ -16,7 +16,7 @@ MainWindow::MainWindow(QWidget *parent) :
 {
     ui->setupUi(this);
 
-    setWindowTitle("Magic Fountain Alpha " + GLOBAL_VERSION);
+    setWindowTitle("MagicFountain " + GLOBAL_VERSION);
 
     int fontId = QFontDatabase::addApplicationFont(":/fonts/Courier Prime.ttf");
 
@@ -32,27 +32,27 @@ MainWindow::MainWindow(QWidget *parent) :
     m_settings = new QSettings("MagicFountain", "MagicFountain");
 
     if (m_settings->contains("language")) {
-        QString language = m_settings->value("language").toString();
-        QLocale locale(language);
+        m_language = m_settings->value("language").toString();
+        QLocale locale(m_language);
 
         if (m_qtTranslator.load("qt_" + locale.name(),
                             QLibraryInfo::location(QLibraryInfo::TranslationsPath))) {
             qApp->installTranslator(&m_qtTranslator);
         }
 
-        if (m_translator.load(QString("magicfountain_") + language, QString(":/locales/"))) {
+        if (m_translator.load(QString("magicfountain_") + m_language, QString(":/locales/"))) {
             qApp->installTranslator(&m_translator);
         }
     } else {
         QLocale locale = QLocale::system();
-        QString language = QLocale::languageToString(locale.language());
+        m_language = QLocale::languageToString(locale.language());
 
         if (m_qtTranslator.load("qt_" + QLocale::system().name(),
                             QLibraryInfo::location(QLibraryInfo::TranslationsPath))) {
             qApp->installTranslator(&m_qtTranslator);
         }
 
-        if (m_translator.load(QString("magicfountain_") + language, QString(":/locales/"))) {
+        if (m_translator.load(QString("magicfountain_") + m_language, QString(":/locales/"))) {
             qApp->installTranslator(&m_translator);
         }
     }
@@ -64,6 +64,7 @@ MainWindow::MainWindow(QWidget *parent) :
     ui->textBrowser_preview->setFont(courierfont);
 
     connect(ui->plainTextEdit_fountaineditor, SIGNAL(textChanged()), this, SLOT(refreshPreview()));
+    connect(ui->plainTextEdit_fountaineditor, SIGNAL(modificationChanged(bool)), this, SLOT(refreshTitleBar(bool)));
     connect(ui->actionExport_as_PDF, SIGNAL(triggered()), this, SLOT(exportAsPDF()));
     connect(ui->actionExport_as_HTML, SIGNAL(triggered()), this, SLOT(exportAsHTML()));
     connect(ui->actionNew, SIGNAL(triggered()), this, SLOT(newFile()));
@@ -98,7 +99,32 @@ MainWindow::~MainWindow()
     delete ui;
 }
 
-void MainWindow::refreshPreview() {
+void MainWindow::closeEvent( QCloseEvent *event )
+{
+    event->ignore();
+
+    slot_checkAndSaveScript();
+
+    event->accept();
+}
+
+void MainWindow::refreshTitleBar(bool modified)
+{
+    QString title = "MagicFountain " + GLOBAL_VERSION;
+
+    if (!filepath.isEmpty()) {
+        title.append(" " + filepath);
+    }
+
+    if (modified) {
+        title.append(" *");
+    }
+
+    setWindowTitle(title);
+}
+
+void MainWindow::refreshPreview()
+{
     if (currentScript != nullptr) {
         delete currentScript;
     }
@@ -184,25 +210,39 @@ void MainWindow::print() {
 }
 
 void MainWindow::newFile() {
-    QString filename = ":/data/default.fountain";
-    if(filename.isEmpty()){
-        return;
-    }
+    slot_checkAndSaveScript();
+
+    QString filename = ":/data/default_" + m_language + ".fountain";
 
     QFile file(filename);
     if (file.open(QIODevice::ReadOnly | QIODevice::Text)) {
-        ui->plainTextEdit_fountaineditor->clear();
-        ui->plainTextEdit_fountaineditor->appendPlainText(file.readAll());
         filepath.clear();
+        ui->plainTextEdit_fountaineditor->setPlainText(file.readAll());
+        ui->plainTextEdit_fountaineditor->document()->setModified(false);
         file.close();
 
         QTextCursor cursor = ui->plainTextEdit_fountaineditor->textCursor();
         cursor.movePosition(QTextCursor::Start);
         ui->plainTextEdit_fountaineditor->setTextCursor(cursor);
+    } else {
+        file.setFileName(":/data/default_en.fountain");
+
+        if (file.open(QIODevice::ReadOnly | QIODevice::Text)) {
+            ui->plainTextEdit_fountaineditor->setPlainText(file.readAll());
+            ui->plainTextEdit_fountaineditor->document()->setModified(false);
+            filepath.clear();
+            file.close();
+
+            QTextCursor cursor = ui->plainTextEdit_fountaineditor->textCursor();
+            cursor.movePosition(QTextCursor::Start);
+            ui->plainTextEdit_fountaineditor->setTextCursor(cursor);
+        }
     }
 }
 
 void MainWindow::openFile() {
+    slot_checkAndSaveScript();
+
     QString filename = QFileDialog::getOpenFileName(this,
                                                     tr("Open Fountain file"),
                                                     QStandardPaths::standardLocations(QStandardPaths::DocumentsLocation).first(),
@@ -213,9 +253,9 @@ void MainWindow::openFile() {
 
     QFile file(filename);
     if (file.open(QIODevice::ReadOnly | QIODevice::Text)) {
-        ui->plainTextEdit_fountaineditor->clear();
-        ui->plainTextEdit_fountaineditor->appendPlainText(file.readAll());
         filepath = filename;
+        ui->plainTextEdit_fountaineditor->setPlainText(file.readAll());
+        ui->plainTextEdit_fountaineditor->document()->setModified(false);
         file.close();
 
         QTextCursor cursor = ui->plainTextEdit_fountaineditor->textCursor();
@@ -239,6 +279,7 @@ void MainWindow::saveAs() {
         stream << ui->plainTextEdit_fountaineditor->toPlainText();
         file.close();
         filepath = filename;
+        ui->plainTextEdit_fountaineditor->document()->setModified(false);
     }
 }
 
@@ -253,6 +294,7 @@ void MainWindow::quickSave() {
         QTextStream stream(&file);
         stream << ui->plainTextEdit_fountaineditor->toPlainText();
         file.close();
+        ui->plainTextEdit_fountaineditor->document()->setModified(false);
     }
 }
 
@@ -305,17 +347,17 @@ void MainWindow::slot_actionFountain_Syntax()
 
 void MainWindow::slot_actionLanguage(QAction *action)
 {
-    QString language = action->iconText();
-    QLocale locale(language);
+    m_language = action->iconText();
+    QLocale locale(m_language);
 
     if (m_qtTranslator.load("qt_" + locale.name(),
                         QLibraryInfo::location(QLibraryInfo::TranslationsPath))) {
         qApp->installTranslator(&m_qtTranslator);
     }
 
-    if (m_translator.load(QString("magicfountain_") + language, QString(":/locales/"))) {
+    if (m_translator.load(QString("magicfountain_") + m_language, QString(":/locales/"))) {
         qApp->installTranslator(&m_translator);
-        m_settings->setValue("language", language);
+        m_settings->setValue("language", m_language);
     }
 }
 
@@ -358,4 +400,13 @@ void MainWindow::changeEvent(QEvent *event)
     }
 
     QMainWindow::changeEvent(event);
+}
+
+void MainWindow::slot_checkAndSaveScript()
+{
+    if (currentScript != nullptr && ui->plainTextEdit_fountaineditor->document()->isModified()) {
+        if (QMessageBox::question(this, QString(tr("Save Fountain file")), QString(tr("Do you want to save current script ?"))) == QMessageBox::Yes) {
+            saveAs();
+        }
+    }
 }